gateway: allow plaintext auth for localhost
"""Given that there are unfortunately still Zarafa systems around using
saslauthd without pam_mapi but rimap instead, the
"disable_plaintext_auth" feature prevents them from enabling this
option as rimap does not support SSL/TLS;
https://jira.zarafa.com/browse/ZCP-12473 contains an example report by
a Zarafa customer. Thus, this patch adds an exception if the source
IPv4 address is "127.0.0.1" and allows, even if
"disable_plaintext_auth" is enabled, a cleartext authentication. It was
a design decision to check only for 127.0.0.1/32 rather 127.0.0.0/8
because there seem to be systems where the loopback network except
127.0.0.1/32 is routable?!""" -rsc

rimap, a software which was last touched in 2003 according to
Sourceforge file timestamps, does not support TLS. To make it run
anyway, enable plaintext authentication for local connections. Testing
for "127.0.0.1" is error-prone (fails to match "::ffff:127.0.0.1") and
does not support other local addresses. Just ask the kernel via
ECChannel::peer_is_local for network namespace locality, as that also
covers IPv6 and things like 192.168.1.1. -jengelh

References: ZCP-12473