common: replace broken three-way comparison
MSVC warning that hinted towards the problem:

4>.\Util.cpp(1184) : warning C4244: '=' :
  conversion from 'float' to 'int', possible loss of data
4>.\Util.cpp(1195) : warning C4244: '=' :
  conversion from 'double' to 'int', possible loss of data

Computing a minus b is a problem, _especially_ for floating point, but
also for all other cases.

Consider a = 1.0, b = 0.75. The difference, 0.25, will be truncated to
0, which means equal, but a and b are not.

For other carefully chosen a and b, a minus b will lead to an
underflow, and also wrong sorting order.

References: ZCP-13427