inetmapi, dagent, gateway: make use of the librosie sanitizer
The HTML in MIME parts may contribute to cross-site scripting issues,
so it is probably best filtered. WebApp already feeds it through the
Squirrelmail sanitizer, but we would like to also offer the filtering
in a more central place such that IMAP and MAPI users can benefit as
well.

References: ZCP-13472