installer: resolve perl 5.22 fatal errors
Starting with Perl 5.22, "defined(@array)" is now rejected.

$ ../installer/ldap/ol-schema-migrate.pl -b \
  ../installer/ldap/zarafa.schema > ../installer/zarafa.ldif
Can't use 'defined(@array)' (Maybe you should just omit the
defined()?) at ../installer/ldap/ol-schema-migrate.pl line 151.

Changing if (defined(@a)) to if (@a) is equivalent to
if (scalar(@a) != 0). If the array has zero elements however, we do
not need to explicitly check for it, but just do the += operation.

References: ZCP-13595