zarafa-server: be pessimistic about daemon startup
Only return with true if we managed to get to the main loop. Otherwise,
return with an error. This fixes lots of situations where the daemon
claimed to start but did not (e.g. when not being able to connect
to the database or to write to the attachment folder).