build: remove -Wno-vla
"-Wno-vla" is not recognized by Collax's gcc 4.2.4.

VLAs are a GNU extension, and this is why their use incurs a compiler
warning, and subsequently a flag to turn off that warning.

There appear to be no VLAs in use in the code anyway, and removing the
flag will not produce a new warning either in gcc 4.8.