build: rename directories in ECtools/
We would like to try moving to a non-recursive build within the
ECtools/ subtree to keep all CPU cores busy at all times, and reducing
build times. In that move, it seems desirable to have the binaries be
right in ECtools/ rather than the individual source subdirectories. To
facilitate this, the subdirectory must not carry the same name as the
executable.