build: remove useless NOTPARALLEL statement from Makefile
The prerequisites of the special .NOTPARALLEL target have no meaning,
they are ignored. Targets, such as install-exec-hook, are only
executed once anyway.

Remove NOTPARALLEL, which caused the swig/ subdirectory to build
longer than it ought to on a multi-core system.

On i7-2600K make -j8, it goes from 28.7s to 10.6s.