python-zarafa: overload __eq__/__ne__ for Store/User/Folder
Overload __eq__ and __ne__ for Store, User, and Folder, and use them
in zarafa-search.

we want to avoid indexing junk folders and outboxes. To have a simple
check, we need to overload equality/inequality operators for the
Folder class. While at it, I implemented them for User and Store right
away. Still some improvements to be made there, but in most cases, this
should work.

References: ZCP-13632