build: move uuid.h requirement out of platform.linux.h
Move #include uuid.h out of platform.linux.h, it is not used anywhere
else, and only forces zcp-nonoss.spec to BuildRequire it even though
it is not used there either.

As a result, some source files that implicitly relied on the
gettimeofday function prototype started tripping. Fix them up.