zcp_common_util: revert making ECLocale a class
Revert SVN commit 50269 for ECLocale class definition.
This concludes the partial revert already done in SVN commit 50272.

The mangled-name function signatures can already change between
different architectures, so it is hard to write a libmapi.sym with
full function parameter specification. When using wildcards for the
parameter list however, we do not need to care about the types
anymore, and therefore, ECLocale can go back to being the simple
typedef it was.