build: provide namespace-sanitized config.h to API users
Users of the ZCP API (zcp-nonoss, zstatsd) would, for example,
determine whether or not to use the ICU-enabled or ICU-less variant of
the ECLocale type. However, HAVE_ICU is nowhere defined in the public
ZCP API, which means downstream packages have to rely on assumptions
what the API looks like.

Have ZCP provide a header file, zconfig.h, which describes its own
compiled state, so that downstream packages do not run risk of
erroneously picking the wrong ECLocale variant.