build: replace icu-config call with pkg-config
On Debian 7, `icu-config --cflags` contains flags other than -D and
-I, causing @icu_CFLAGS@ in common/zarafa.pc(.in) to hold flags like
-Wwhatever. The Cflags: line in zarafa.pc is used in zcp-nonoss
swig/python/Makefile.am as ${zcp_CFLAGS}, and this is also passed to
swig, but swig does not know gcc-specific flags.

The short fix would be to use `icu-config --cppflags`, but given that
icu-config itself is deprecated and the manual recommends using
pkgconfig for it, this commit replaces icu.m4 with the appropriate
pkgconfig calls and Makefile.am edits.