dagent: support for SO_BINDTODEVICE
When using socket activation as e.g. supported by systemd, one can
use a BindToDevice= directive to limit connections to those which
use a L3 destination address on the specified interface (IPv6), or
any L3 destination address that is considered local (IPv4).

Add support for device binding even in the absence of systemd socket
activation. This makes it possible to offer the dagent service on both
IPv6 and IPv4 over a single socket file descriptor (for which the
wildcard binding address needs to be used) without making service
reachable from outside the machine even in the absence of a local
packet filter.