installer: create and chown /var/run/zarafad at start
For legacy sysvinit systems, the init script(s) will now create
/var/run/zarafad and chown it accordingly. For systemd environments,
this is taken care of through a tmpfiles.d(5) snippet, which is
activated by systemd on boot, or by the %post section of the distro
package.

The directory name "/var/run/zarafad" is chosen because
"/var/run/zarafa" is already occupied by the zarafa-server socket,
even if the overlap is just for a transitional period.

Matching user creation logic also is to occur with %post and/or
sysusers.d(5), for which a snippet is provided as well.

References: ZCP-13329