server: emit message on early termination
When, for example, the database access configuration is incorrect,
`zarafa-server -F` terminates silently. When -F is used, it should emit
a message to the console, not only in the log file.

This problem was solved by:

When the (parent) process terminates, it checks for er != erSuccess and
then emits a helpful message that the user should check the logfile.

References: ZCP-12490