common: avoid late reinvocation of PrivatePipe signal handlers
For the non-threaded case, the PrivatePipe::PipePassLoop function did
not reset the signal handlers it installed. If a SIGHUP arrives just
after "delete lpFileLogger", the PrivatePipe::sighup function may be
invoked. Given PrivatePipe::m_lpFileLogger is a dangling pointer since
the delete, this would lead to undefined behavior - that is to say,
usually a crash in this case.

Remedy this by resetting the signal handlers before releasing the
logger.