pyza: drop privs even in foreground mode when so requested
When zarafa daemons utilizing the pyza daemonize() function are started
with their foregrounding option, they fail to switch to alternate users
as defined in the "run_as_user" and "run_as_group" parameters in the
config file. The cause for this is that no daemon context is created
for the options.foreground case.

This affects zarafa-search and, from zcp-nonoss, zarafa-presence.

Do unconditionally create the context, and set c.detach_process
appropriately.

References: ZCP-13712