dagent: change server_bind_intf default to "all"
When dagent is started by the system, this happens as root, and the
process is able to execute the SO_BINDTODEVICE setsockopt, before it
changes to the unprivileged user.

When however starting dagent from an already unprivileged context
(lacking CAP_NET_ADMIN), SO_BINDTODEVICE will not succeed. As this mode
of operation is popular with developers, redact the server_bind_intf
default.