pyza: Add MVPropMap support to Python-MAPI
The MVPropMap only holds the enabled/disabled features at the moment,
and is implemented in Python-MAPI as a list of MVPropMaps which
contains a ulPropId and a list of Values. The Values are only strings
at the moment.

This resolves the bug where the enabled/disabled features are reset
when we call SetUser() in Python-Zarafa without passing the sMVPropMAp
struct.

This also improves the unittest, since they now shell out to
enable/disable the user features.

Introduce in Python-Zarafa User.features, which shows the current
enabled features, user.{add_feature,remove_feature} can be used as
helpers. For older Python-MAPI versions, we will throw a
ZarafaNotSupported error.