installer: try two pid locations for killing via sysvinit
When the package update is installed, the sysvinit script is replaced
with a new variant that looks for PID files in the new location,
/var/run/zarafad/$xyz.pid. However, the still-running daemon has its
process number written to the old location. As a result, the daemon
will not be stopped.

There is no reliable fix possible here other than dropping the sysvinit
sucker and using some init system which remembers the PIDs of its
(first-level) children itself.

References: ZCP-13731