dagent: better error reporting on duplicate email addresses
If two accounts have the same email address, dagent will reject a mail
with "503 5.1.1 User does not exist (in reply to RCPT TO command)".
Rejecting the mail is fine, but the status code is wrong. It should be
"5.1.4 Destination mailbox address ambiguous".

Solved by letting ResolveUser(s) export the resolve flags to the
caller. This way, it can distinguish between not found and ambiguous.

References: ZCP-12356