inetmapi: allow disabling the heuristic charset reinterpretation
Trying multiple character sets was introduced in SVN r50749 and was an
attempt to understand mails which had a "good" character set
declaration in the HTML <meta> tag, and a "bad" one in Content-Type —
or vice-versa.

If there was just one unconvertible sequence in the input sequence,
(perhaps as a result of a bit error, or because another mailer mangled
it), VMIMEToMAPI would try conversion with the 2nd choice character
set. That result may be totally worse than ignoring errors with the
1st choice character set.

Provide a knob to disable the heuristic character set search, and make
it the default too.