inetmapi: do not ignore parts in an alternative container
When the findBestAlternative function encounters a part without any
Content-Type header, it ignores it. This is wrong for two reasons:

1. If such an untagged part is the only one in an alternative
container, the function would return a list with zero elements, causing
the parent function, dissect_multipart, to not execute the for loop and
not select any alternative, essentially dropping the message part.

2. The absence of the Content-Type header is not a reason to ignore an
alternative. RFC 2046 §5.1 ¶2: """To begin with, NO header fields are
actually required in body parts. A body part that starts with a blank
line, therefore, is allowed and is a body part for which all default
values are to be assumed. In such a case, the absence of a Content-Type
header usually indicates that the corresponding body has a content-type
of "text/plain; charset=US-ASCII"."""

Fix that by pushing to lBodies even if no Content-Type was found.

Update the function commentary while at it, and make it emit unsigned
ints for good practice (attachment numbers cannot be negative).

NOTE: Behavioral change. This commit essentially reverts the supposed
"fix" for ZCP-11581.

References: ZCP-11581.
Part on a series to make ZCP RFC conformant (instigated by ZCP-11713).