inetmapi: throw out pointless feed-re-extract cycle for HTML charset
The parameter_not_found exception is only thrown if the "charset"
variable was of length zero. That would have caused the create()
function call to ignore the charset= parameter.

Therefore, the whole convoluted try-catch block can go to limbo,
because stowing the charset in a vmime::headerField just to extract it
again is silly. Just pass the character set as found in the HTML
directly into getCompatibleCharset. (And take care of the zero length
case.)