pyza: avoid "Terminating on signal 15" when worker instance terminates
backup-plus, for some reason, on Debian 8 at least, currently shows the
message "Terminating on signal 15" for each worker instance when
backup-plus finishes. The main process still exits with error code 0,
so the messages are just annoying, and can be ignored.

Yeah, backup-plus could probably terminate workers explicitly in some
way, but this should not be necessary.