inetmapi: avoid sending trailing \0 into PR_BODY_W attribute
c_str() returns a pointer to a region which is always \0-terminated,
but there is no apparent reason to do this. (Were the body a PT_BINARY,
this would be outright wrong.) For text, libmapi appears to recompute
the body length anyway, as the \0 does not end up in the database.
So we might just as well drop that.