pyza: make Service and Worker classes ignore SIGHUP
Both when reloading a service via an init script, and logrotating as
setup by default in /etc/logrotate.d/zarafa (zarafa-common), a HUP
signal is used to indicate a "reload" situation. As we do not support
reloading in python-zarafa at the moment, let's just ignore the signal
for now. Otherwise, python-zarafa-based services such as search and
presence get killed at logrotation time. Not sure why we have not
noticed the problem earlier.

References: ZCP-13788