daemons: chown PID file and socket to user running the service
zarafa-server would first create the sockets, and then change to the
other user ID. This would mean other services like zarafa-licensed
would be unable to connect. (The PID files also created as root, but
this is not so important, since they are not needed for operation. But
it gets fixed anyway.)

The Python-based programs already seem to be exhibiting the new
behavior.