libserver: remember up to 10 ICS states rather than 3
We currently support remembering the last 3 states (current one plus 2
old ones) to support some kind of rollback functionality/re-exporting
of old states. As described in the code, this number is pretty
arbitrary.

Outlook over ActiveSync apparently likes to go back further in time, so
let's try to remember up to 10 states for now. Note that this will
probably multiply the size of the "syncedmessages" table.

References: ZCP-13757