search: extract "indexing" terms from search text
When indexing, the backend search engine will strip special characters
in a certain way to generate the final search "terms" which are stored
in the DB. We want to use the same approach when searching (so queries
containing special characters), or there will be a mismatch. Therefore,
we add another method to the plugin interface for this. We would like
to hide this stripping fully, but unfortunately, we still need to add
"*" after each term.

References: ZCP-13784