search: limit search terms to 8
Currently, it is possible to copy-and-paste a Wikipedia page into the
WebApp search box, causing havoc and denial-of-service.

Clients should probably limit this, but to be safe, the backend now also
limits the total number of search terms to 8. Quote me on this:

"8 search terms should be enough for everybody.."

References: WA-9167