daemons: ensure MAPIInitialize comes after StartLoggerProcess
When logging to a file, StartLoggerProcess would decref/replace the
logger object and hand out a pointer to a new one. This means that all
other ref holders (MAPIInitialize > HrCreateM4LServices calls AddRef)
continue to work with the original ECLogger_File object — which cannot
be intentional, because it could interleave writes with the new
ECLogger_Pipe writer.

Call MAPIInitialize after StartLoggerProcess.

References: ZCP-13816