dagent: allow empty Out Of Office messages
dagent's behavior to refuse sending OOF notifications with an empty
body is arbitrary and harmful. The user is not notified about the
refusal, also because the zarafa-server MAPI happily accepts property
combinations that dagent would not like.

There is nothing wrong with empty bodies. Just allow them already.

References: ZCP-13729