m4lcommon: push default socket path one layer down
Change HrOpenECSession such that it accepts a NULL socket path or an
empty string and will automatically fill it in before calling SOAP.
This way, MAPI clients do not have to call GetServerUnixPath(), but
leave everything to the implementation.

In fact, since GetServerUnixPath is not exposed via php-ext, WebApp has
had no way but to hardcode the location (leading to its share of
problems).

References: ZCP-13856