server: reject compression requests when peer is local
When the remote peer in a connection is local as well, there is hardly
a justification for compression, so let's disable it for that case.

It turns out that disabling compression on the server side is easier
than finding the place ten layers deep where to do it for the client
side. Make it so!

References: ZCP-13667