pyza: fix search crash for store without junk/outbox folders
For a user store, it is unusual, but for a public store, it is
normal to not have a junk/outbox/etc. folder..

Change python-zarafa to return None when such a folder is requested,
instead of raising an exception. This is consistent with how we treat
similar situations elsewhere in the code. Exceptions are only for
really exceptional cases, such as when a folder with a explicit name is
requested.