pyza: split folder() key arg into name and entryid
incompatible change, to avoid naming conflicts between folder
names and entryids. suppose we have a folder with as name an
entryid of another folder, and try to look for a folder with
this name/entryid. not that there's a large chance that this
will actually happen, but we don't want to leave the possibility
open.

Conflicts:
	swig/python/zarafa/zarafa/__init__.py