inetmapi: retain attachment status even for top-level body part
VMIME unwraps single attachment multiparts. It even has a comment in
its source code about that behavior: """If there is only one part in
the message, move it into the message (hence, the message will not be
multipart...)""". The RFC2822 messages so emitted looks acceptable and
check out visually in other MUAs, including Outlook, Alpine and
Thunderbird.

However, when this RFC2822 message is parsed back into a MAPI message,
the attachment status gets lost, and it will not show up in WebApp as
an attachment. In addition, zarafa-gateway (if made not to use the full
copy stored in the PR_EC_IMAP_BODY property) shows the
transmogrification as well.

Fix: Do not special case top-level disposition status.

References: ZCP-13851