libserver: resolve coverity remark about NULL deref
6. var_compare_op: Comparing this->lpData to null implies that
this->lpData might be null.
CID 158780 (#1 of 1): Dereference after null check (FORWARD_NULL)
7. var_deref_model: Passing null pointer other.lpData to memcmp,
which dereferences it.
CID 158781 (#1 of 1): Dereference after null check (FORWARD_NULL)
7. var_deref_model: Passing null pointer this->lpData to memcmp,
which dereferences it.

There are four possible cases.

1. lpData == NULL && other.lpData != NULL
	==> already handled by return true
2. lpData != NULL && other.lpData == NULL
	==> already handled by return false
3. lpData != NULL && other.lpData != NULL
	==> is to lead to memcmp
4. lpData == NULL && other.lpData == NULL

Normally, cData is 0 when lpData == NULL and therefore, memcmp will
never attempt to dereference lpData, but of course, there is no
programmatical guarantee for it. This is why Coverity warns.

Catch case 4 with an extra check.

References: ZCP-12524