search: make 'syncrun' sleep longer
the testset may ask search to finish syncing by sending a SYNCRUN
command. currently this is implemented as a 5 second sleep, which
is assuming that the system is fast enough for search to always
keep up within this time. on a slower system that my own, such as
jenkins, it appears this is not enough. so I changed it to 10
seconds.

the real solution is to use a multiprocessing signal, but let's
see what happens with this first.

ZCP-7269