build: detect unsupported variable expansion in --with-userscript-dir
Commit bp/zcp/7.2.3-539-g56a0614 changed the default value for the
USERSCRIPTDIR variable to include variables. Unfortunately, that only
works if later evaluted by a Makefile, which is not the case for
AC_CONFIG_FILES. Detect variables and just reject them because there
is no point in spending time on it.

References: ZCP-13890