misc: resolve potential memleaks
LLVM's scan-build found some potential memory leaks throughout the
code, although it will be hard to mee these conditions it would be nice
to resolve them.

In ECLicenseClient, we leak memory when lppResponse is NULL, therefore
start allocating when it has a value.

SOAPUtils leaks memory when CopyRestrictTable fails, therefore check
the er coniditon in the exit label and delete the object.

Fileutil leaks memory when mmap fails and we do not read anything
(offset == 0), since we do not set lppBuffer = lpBuffer, the caller
will never free it, which causes a memory leak.