ldap: do not end LDAP searches after one page
On a particular system, the Zarafa user and contact list in LDAP
amounted to some 4000+ users/contacts. The global addressbook would
show a large number (possibly all of them). The SQL database, upon
statistical analysis, showed about 40 ACTIVE_USER, 7 INACTIVE_USER, and
roughly 3900 INACTIVE_CONTACT objects in the "users" table. Despite all
that, `zarafa-admin -l` would only 2 (active) users.

The ldap_page_size option in /etc/zarafa/ldap.cfg was set to 1000. The
LDAPUserPlugin::getAllObjectsByFilter function would initiate a paged
readout, but end the loop after just one iteration, returning just 1000
entries to the caller (of which 998 happened to be contacts, and 2 were
users).

RFC 2696 specifies that the cookie used for paging is of type "OCTET
STRING" and that "the client MUST consider the cookie to be an opaque
structure and make no assumptions about its internal organization or
value." The last part is what ZCP botched, because it assumed it was a
NUL-terminated C string, which an OCTET STRING is not.

As openldap 2.4 itself tests for bv_val>0 when evaluating more_pages,
the comment must be wrong.

References: ZCP-13771