ZCP-13390: This fixes the out-of-the-box experience with WebAccess to actually allow unsafe logins per default (as we did in the past) by making the php_flags for cookiesecure and httponly per default only available when mod_ssl is really enabled.
This makes default, unconfigured state of webaccess per default insecure, but
any admin nowadays knows that SSL for securing something is the minimal
requirement for having any level of security. So locking down parameters
to make it per-se disfunctional without any error message is no fix.

Since this is like the last webaccess fix at all (leaving support) we should
at least allow the guys to just be able to use it as it has been ever since.

This fixes the only partial commit of r47765, which helped ubuntu, but killed RHEL and SLES out-of-the-box. This commit fixes all distributions.