common: avoid overshooting nlmsg_len
RTA_LENGTH(x) adds to x the sizeof the attribute header. As rta_len
already includes the attribute header, it is incorrect to add that
header size again for nlmsg. So we are sending extra NUL bytes to the
kernel. The nla_ok function in the kernel rightfully returns false and
attribute parsing stops prematurely. Since the remaining bytes are
garbage anyway, it is not a problem, but it leads to the emission of
messages in the kernel ring buffer:

netlink: 4 bytes leftover after parsing attributes in process `abc'.

References: ZCP-13921