caldav: make Lightning workaround specific to Lightning
Cherry-pick KC commit bp/8.0.1-231-gfd2c22b5 .

Apparently, Thunderbird Lightning does not like to receive
calendar-home-set for public folders, so we disabled this in the past
(empty value). A recent Apple client however requires a value for this,
so we specialize the Lightning hack to only apply for Lightning.

References: KC-145