inetmapi: set calendar properties of main message
Cherry-pick KC commit bp/8.2.0-508-g3811103b .

When an e-mail is received which has both a text/html part and an ICS
part, then the ICS object would be stored as an attachment so that
neither the HTML body nor the ICS DESCRIPTION field get lost in any
way. The MAPI properties for calendaring would then be set on the
attachment, but they are really needed on the main message instead. Set
them there, that is, of course, except for the body if one was already
present.

References: KC-138